Export reset_config in nucliadb package
