--- test-requirements.txt
@@ -1,19 +1,0 @@
-mypy>=1.8
-pytest==7.4.2
-pytest-asyncio~=0.18.0
-pytest-cov==3.0.0
-types-setuptools
-types-redis
-httpx>=0.23.0
-pytest-docker-fixtures>=1.3.17
-pytest-benchmark
-docker>=7.1.0
-grpcio-health-checking==1.44.0
-requests
-maturin
-pytest-lazy-fixtures
-pytest-mock
-grpc-stubs>=1.44.0
-aioresponses
-# required for pytest-docker-fixtures
-psycopg2-binary
